feat: bring the all friends under a group called 'all' [wip]
